+++ libs/form-component/src/app-components/Card/Card.tsx
@@ -1,0 +1,52 @@
+import React from 'react';
+
+import { Card, Heading, Paragraph } from '@digdir/designsystemet-react';
+
+import classes from './Card.module.css';
+
+export type AppCardProps = {
+  title?: React.ReactNode;
+  description?: React.ReactNode;
+  footer?: React.ReactNode;
+  media?: React.ReactNode;
+  mediaPosition?: 'top' | 'bottom';
+  color?: Parameters<typeof Card>[0]['color'];
+  children?: React.ReactNode;
+  variant?: 'tinted' | 'default';
+  className?: string;
+  ref?: React.Ref<HTMLDivElement>;
+};
+
+export function AppCard({
+  title,
+  description,
+  footer,
+  media,
+  color = 'neutral',
+  mediaPosition = 'top',
+  children,
+  variant = 'tinted',
+  className,
+  ref,
+}: AppCardProps) {
+  return (
+    <Card data-color={color} variant={variant} className={className} ref={ref}>
+      {media && mediaPosition === 'top' && (
+        <Card.Block className={classes.mediaCard}>{media}</Card.Block>
+      )}
+      <Card.Block>
+        {title && <Heading data-size='md'>{title}</Heading>}
+        {description && <Paragraph>{description}</Paragraph>}
+        {children}
+      </Card.Block>
+      {footer && (
+        <Card.Block>
+          <Paragraph data-size='sm'>{footer}</Paragraph>
+        </Card.Block>
+      )}
+      {media && mediaPosition === 'bottom' && (
+        <Card.Block className={classes.mediaCard}>{media}</Card.Block>
+      )}
+    </Card>
+  );
+}
