+++ libs/form-component/src/app-components/Card/index.ts
@@ -1,0 +1,2 @@
+export { AppCard } from './Card';
+export type { AppCardProps } from './Card';
