+++ libs/form-component/src/app-components/index.ts
@@ -1,0 +1,1 @@
+export * from './Card';
