--- src/App/frontend/src/app-components/Card/Card.tsx
@@ -1,68 +1,0 @@
-import React from 'react';
-
-import { Card, Heading, Paragraph } from '@digdir/designsystemet-react';
-
-import { useTranslation } from 'src/app-components/AppComponentsProvider';
-import classes from 'src/app-components/Card/Card.module.css';
-import type { TranslationKey } from 'src/app-components/types';
-
-type AppCardProps = {
-  title?: TranslationKey;
-  description?: TranslationKey;
-  footer?: TranslationKey;
-  media?: React.ReactNode;
-  mediaPosition?: 'top' | 'bottom';
-  color?: Parameters<typeof Card>[0]['color'];
-  children?: React.ReactNode;
-  variant?: 'tinted' | 'default';
-  className?: string;
-  ref?: React.Ref<HTMLDivElement>;
-};
-
-export function AppCard({
-  title,
-  description,
-  footer,
-  media,
-  color = 'neutral',
-  mediaPosition = 'top',
-  children,
-  variant = 'tinted',
-  className,
-  ref,
-}: AppCardProps) {
-  const { TranslateComponent } = useTranslation();
-
-  return (
-    <Card
-      data-color={color}
-      variant={variant}
-      className={className}
-      ref={ref}
-    >
-      {media && mediaPosition === 'top' && <Card.Block className={classes.mediaCard}>{media}</Card.Block>}
-      <Card.Block>
-        {title && (
-          <Heading data-size='md'>
-            <TranslateComponent tKey={title} />
-          </Heading>
-        )}
-        {description && (
-          <Paragraph>
-            <TranslateComponent tKey={description} />
-          </Paragraph>
-        )}
-
-        {children}
-      </Card.Block>
-      {footer && (
-        <Card.Block>
-          <Paragraph data-size='sm'>
-            <TranslateComponent tKey={footer} />
-          </Paragraph>
-        </Card.Block>
-      )}
-      {media && mediaPosition === 'bottom' && <Card.Block className={classes.mediaCard}>{media}</Card.Block>}
-    </Card>
-  );
-}
